+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 
 - Added support for applying Snowflake Cortex functions `Summarize` and `Sentiment`.
 - Added support for `Series.hist`.
+- Added support for list values in `Series.str.get`.
 
 ### Snowpark Local Testing Updates