+++ CHANGELOG.md
@@ -37,6 +37,13 @@
 
 - Fixed a bug in Dataframe.join that caused columns to have incorrect typing.
 
+### Snowpark Local Testing Updates
+
+#### Bug Fixes
+
+- Fixed a bug in when statements that caused incorrect results in the otherwise clause.
+
+
 ## 1.27.0 (2025-02-03)
 
 ### Snowpark Python API Updates