+++ CHANGELOG.md
@@ -30,6 +30,12 @@
 - Added support for applying Snowflake Cortex functions `Summarize` and `Sentiment`.
 - Added support for `Series.hist`.
 
+### Snowpark Local Testing Updates
+
+#### New Features
+
+- Added support for NULL_IF parameter to csv reader
+
 ## 1.27.0 (2025-02-03)
 
 ### Snowpark Python API Updates