+++ CHANGELOG.md
@@ -34,7 +34,8 @@
 
 #### New Features
 
-- Added support for NULL_IF parameter to csv reader
+- Added support for NULL_IF parameter to csv reader.
+- Added support for `date_format`, `datetime_format`, and `timestamp_format` options when loading csvs.
 
 ## 1.27.0 (2025-02-03)